+++ Sources/UI/Style/ChatViewStyle.swift
@@ -19,6 +19,8 @@ public struct ChatViewStyle {
     public var incomingMessage: MessageViewStyle
     /// An outgoing message view style.
     public var outgoingMessage: MessageViewStyle
+    /// Style for AvatarView to be displayed in the Navigation Right Bar Button Item.
+    public var avatarViewStyle: AvatarViewStyle
     
     /// The default chat view style (dynamic for iOS 13+).
     public static let `default`: ChatViewStyle = {
@@ -42,11 +44,13 @@ public struct ChatViewStyle {
                 outgoingMessage: MessageViewStyle = .init(alignment: .right,
                                                           backgroundColor: .chatSuperLightGray,
                                                           borderWidth: 0,
-                                                          reactionViewStyle: .init(alignment: .right))) {
+                                                          reactionViewStyle: .init(alignment: .right)),
+                avatarViewStyle: AvatarViewStyle = .init()) {
         self.channel = channel
         self.composer = composer
         self.incomingMessage = incomingMessage
         self.outgoingMessage = outgoingMessage
+        self.avatarViewStyle = avatarViewStyle
     }
 }